+++ lang/java/ipc/src/test/java/org/apache/avro/specific/TestSpecificRecordBuilder.java
@@ -47,7 +47,7 @@ public class TestSpecificRecordBuilder {
     Assert.assertTrue(builder.hasName());
     Assert.assertEquals("James Gosling", builder.getName());
     Assert.assertTrue(builder.hasYearOfBirth());
-    Assert.assertEquals(1955, builder.getYearOfBirth());
+    Assert.assertEquals(1955, (int) builder.getYearOfBirth());
     Assert.assertFalse(builder.hasCountry());
     Assert.assertNull(builder.getCountry());
     Assert.assertTrue(builder.hasState());
@@ -59,7 +59,7 @@ public class TestSpecificRecordBuilder {
 
     Person person = builder.build();
     Assert.assertEquals("James Gosling", person.getName());
-    Assert.assertEquals(1955, person.getYearOfBirth());
+    Assert.assertEquals(1955, (int) person.getYearOfBirth());
     Assert.assertEquals("US", person.getCountry());  // country should default to "US"
     Assert.assertEquals("CA", person.getState());
     Assert.assertNotNull(person.getFriends());  // friends should default to an empty list
@@ -75,7 +75,7 @@ public class TestSpecificRecordBuilder {
 
     Person.Builder builderCopy = Person.newBuilder(person);
     Assert.assertEquals("James Gosling", builderCopy.getName());
-    Assert.assertEquals(1955, builderCopy.getYearOfBirth());
+    Assert.assertEquals(1955, (int) builderCopy.getYearOfBirth());
     Assert.assertEquals("US", builderCopy.getCountry());  // country should default to "US"
     Assert.assertEquals("CA", builderCopy.getState());
     Assert.assertNotNull(builderCopy.getFriends());  // friends should default to an empty list
@@ -102,13 +102,13 @@ public class TestSpecificRecordBuilder {
           .setProduct(product)
           .build())
       .build();
-    Assert.assertEquals(datetime, p.getDatetime());
+    Assert.assertEquals(datetime, (long) p.getDatetime());
     Assert.assertEquals(ProductPage.class, p.getPageContext().getClass());
     Assert.assertEquals(product, ((ProductPage)p.getPageContext()).getProduct());
 
     PageView p2 = PageView.newBuilder(p).build();
 
-    Assert.assertEquals(datetime, p2.getDatetime());
+    Assert.assertEquals(datetime, (long) p2.getDatetime());
     Assert.assertEquals(ProductPage.class, p2.getPageContext().getClass());
     Assert.assertEquals(product, ((ProductPage)p2.getPageContext()).getProduct());
